requirements.txt
langchain==0.2.16 
langgraph==0.2.14 
chromadb==0.5.5 
sentence-transformers==3.0.1 
streamlit==1.39.0 
tavily-python==0.5.0 
pypdf==4.3.1 
python-dotenv==1.0.1 
tesseract
gtts==2.5.3 
numpy==1.26.4 
ollama==0.3.2